[Move #108] assets 폴더 위치 변경 완료
