+++ components/layout/Footer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import FooterLogo from "@/app/assets/images/footer_logo.svg";
-import GithubIcon from "@/app/assets/images/icon-github.png";
+import FooterLogo from "@/assets/images/footer_logo.svg";
+import GithubIcon from "@/assets/images/icon-github.png";
 import S from "@/style/commons/commons.module.css";
 import Image from "next/image";
 import Link from "next/link";
