+++ style/register/register.module.css
@@ -13,7 +13,7 @@
   position: absolute;
   width: 100%;
   height: 100vh;
-  background: url("../../app/assets/images/bg.png") center center/cover no-repeat;
+  background: url("../../assets/images/bg.png") center center/cover no-repeat;
   display: flex;
   align-items: center;
   justify-content: center;
